test(agent): split native terminal contracts
