+++ scripts/run-agent-console-contract-tests.mjs
@@ -39,6 +39,14 @@ const result = spawnSync(
       'core',
       'tests',
       'python',
+      'test_agent_native_terminal_contract.py',
+    ),
+    path.join(
+      root,
+      'framework',
+      'core',
+      'tests',
+      'python',
       'test_agent_work_advisory.py',
     ),
     path.join(
